% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TablaAnova.R
\name{TablaAnova}
\alias{TablaAnova}
\title{Tabla ANOVA del DBCA.}
\usage{
TablaAnova(respuesta, tratamiento, bloque, data)
}
\arguments{
\item{respuesta}{(string) nombre de la variable respuesta.}

\item{tratamiento}{(string) nombre de la variable que representa los tratamientos.}

\item{bloque}{(string) nombre de la variable que representa los bloques.}

\item{data}{(\code{data.frame}) Tabla de datos en formato largo con los datos
de los tratamientos, bloques y de la variable respuesta.}
}
\value{
Devuelve una tabla en formato \code{data.frame} con los cálculos correspondientes
   al análisis de varianza.
}
\description{
Obtiene la Tabla de Análisis de Varianza (ANOVA) para un Diseño en Bloques Completamente al Azar (DBCA).
}
\examples{
\dontrun{
## 1 Limpiar la memoria de R
rm(list = ls())

## 2 Cargamos el paquete DBCA
library(DBCA)

## 3 Ruta del archivo y nombre del mismo
archivo <- "C:/Users/Jesus/Documents/datos.csv"

## 4 Lectura de archivo de datos
df <- read.csv(archivo)

## 5 Ejecutamos la función TablaAnova
TablaAnova(respuesta = "y", tratamiento = "trat", bloque = "bloque", data = df)
}
}
